requirements.txt
aiohttp>=3.7.3
fbchat-asyncio>=0.6.21
httpx>=0.16.1
toml>=0.10.2
websockets>=9.1
